ccpa_hackathon_package/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyMuPDF==1.24.10
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
requests==2.32.3
pydantic==2.9.2
huggingface-hub==0.25.1
torch==2.4.1
